Fix coveralls integration
Switch from the deprecated coveralls gem to the Coveralls GitHub
Action.

The coveralls gem fails with SSLError [1] because it sets ssl_version
to TLSv1, and its API endpoint is gone (404 Not Found). There is a
community-maintained replacement called coveralls-ruby-reborn, but the
official integration is through the Coveralls GitHub Action which is
easy enough to use so we just switch to that.

[1] https://www.github.com/lemurheavy/coveralls-ruby/issues/163

Co-authored-by: Tom Levy <tomlevy93@gmail.com>